+++ src/filters/ff_mx.c
@@ -386,7 +386,7 @@ static GF_Err ffmx_process(GF_Filter *filter)
 			if (st->cts_shift) ffpck.pts += st->cts_shift;
 
 			if (ffpck.dts > ffpck.pts) {
-				st->cts_shift = ffpck.dts - ffpck.pts;
+				st->cts_shift = (u32) (ffpck.dts - ffpck.pts);
 				GF_LOG(GF_LOG_INFO, GF_LOG_CONTAINER, ("[FFMux] Negative CTS offset -%d found, adjusting offset\n", st->cts_shift));
 				ffpck.pts = ffpck.dts;
 			}
